backend/requirements.txt
# Cell Collective K-12 Wrapper - Backend Dependencies

# Flask web framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5

# Cell Collective Python SDK
ccapi>=0.1.0

# WebSocket support
python-socketio==5.10.0
python-engineio==4.8.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
black==23.12.1
